2020-7-5 finish college and system manage

+++ src/main/java/com/professions/professions/entity/TbGrade.java
@@ -38,7 +38,7 @@ public class TbGrade implements Serializable {
     /**
      * 学院ID
      */
-    private Integer collegeId;
+    private Long collegeId;
 
     /**
      * 创建时间
